+++ tests/test_property_generation.py
@@ -13,8 +13,8 @@ def test_property_assignment(fz):
     fz.recipe.xml.find("Seeds").attrib['synapseSeed'] = "123"
     fltr = DatasetOperation.initialize(["SynapseProperties"],
                                        fz.recipe,
-                                       None,
-                                       None,
+                                       fz.circuit.source,
+                                       fz.circuit.target,
                                        None)[0]
     data = fltr.apply(fz.circuit)
     have = data.select("pre_gid", "post_gid", "synapseType")
@@ -31,8 +31,8 @@ def test_property_positive_u(fz):
     fz.recipe.xml.find("Seeds").attrib['synapseSeed'] = "123"
     fltr = DatasetOperation.initialize(["SynapseProperties"],
                                        fz.recipe,
-                                       None,
-                                       None,
+                                       fz.circuit.source,
+                                       fz.circuit.target,
                                        None)[0]
     data = fltr.apply(fz.circuit)
     assert data.where('u < 0').count() == 0